web: turn the rootInterface() context into a Lit-Context
This commit converts all calls of the form:

``` JavaScript
rootInterface()?.config?.capabilities.includes(CapabilitiesEnum.CanImpersonate)
```

into:

``` JavaScript
this.can(CapabilitiesEnum.CanImpersonate)
```

It does this via a Mixin, `withCapabilitiesContext`, which understands that these calls
only make sense in the context of a running, fully configured authentik instance, and
that their purpose is to inform authentik components of a user's capabilities.  The
latter is why I don't feel uncomfortable turning a function call into a method; we
should make it explicit that this is a relationship between components.

The mixin has a single single field, `[WCC.capabilitiesConfig]`, where its association with the
upper-level configuration is made. If that syntax looks peculiar to you, good! I've used an explict
unique symbol as the field name; it is inaccessable an innumerable in the object list. The
debugger shows it only as:

```
Symbol(): {
    cacheTimeout: 300
    cacheTimeoutFlows: 300
    cacheTimeoutPolicies: 300
    cacheTimeoutReputation: 300
    capabilities: (5) ['can_save_media', 'can_geo_ip', 'can_impersonate', 'can_debug', 'is_enterprise']
}
```

Since you can't reference it by identity, you can't write to it.  Until every browser supports
actual private fields, this is the best we can do; it does guarantee that field name collisions
are impossible, which is a win.

The mixin takes a second optional boolean; setting this to true will cause any web component using
the mixin to automatically schedule a re-render if the capabilities list changes.

Because this work builds off the work I did to give the Sidebar access to the configuration without
ad-hoc retrieval or prop-drilling, it wasn't necessary to create a new context for it.  That will be
necessary for the following:

TODO:

``` JavaScript
rootInterface()?.uiConfig;
rootInterface()?.tenant;
me();
```