web: re-enable storybook
This is the kind of silly hack that we're required to mess with in
the lovely era of a constantly moving "standard."  The short form
is that Vite, and Storybook-Vite, have adopted a convention of
including the query '?inline' at the end of CSS imports to tell
Vite not to produce TWO css imports-- one for the module and one
for the root.  Unfortunately, this hack doesn't work with the base
rollup bundler that we use, and it's a lot of extra unnecessary
noise in the source code.

This commit adds a script that creates a map from the stock import
string to one that includes the '?inline' annotation, and then
modifies the Storybook configuration to apply that map to the source
code stream during compilation.  It's ugly and hacky, but it does the
job and is fairly well-documented.