This commit begins the decomposition process described in the "Detangling" phase conducted the week of September 8th through the 14th.  In that week, we detangled the UI such that
- Admin did not refer to User
- Elements did not refer to Admin
- Elements did not refer to User
- Common did not refer to Elements
- Components did not refer to Admin

While there is still some cross-referencing that needs to be addressed, that was an excellent start
in saying that there is a hierarchy between the foundation of Contexts (and the API), Applications,
and Elements.

With that hierarchy in mind, this commit enters a whole new phase of the build decomposition,
streamlining, and performance campaign.

The objective is to extract, one by one, components of the system that do not need "awareness" of
other components, and to make them independent, and to create an ecosystem of elements, components,
contexts, verticals (individual user stories, such as "Manage a User" or "Add a Provider"), and
finally whole applications that have both high visibility and legibility, and which create a
dependency graph that will allow developers to work faster, more efficiently, and with greater
confidence.

At least that's the hope.